+++ web-dev/src/Types/Ingredient.ts
@@ -1,5 +1,8 @@
+import { TagType } from "../Forms/CreateIngredientForm";
+
 export interface Ingredient {
   id: number;
   name: string;
   price: number;
+  tag: TagType;
 }