+++ web-dev/src/Types/OptionsMultiSelect.ts
@@ -1,4 +1,7 @@
+import { TagType } from "../Forms/CreateIngredientForm";
+
 export interface OptionsMultiSelectType {
   id: number;
   label: string;
+  tag: TagType;
 }